Add files via upload
The project can be openly-access without the use of any kind of login information. The user can do the inputting of quantity of the ordered items, he/she also need to enter the name of the customer and phone number. After the order is done the user will click the Total button in order to display the total ordered item and click the Generate Bill to display the receipt of the purchased item. This project is useful to those have a grocery business it could help a lot when it comes for automating the calculation of transaction. This project contains the basic functionality of python in order to learn the flow of the system.